A few small fixes.
