+++ src/components/team/MembersList.js
@@ -49,6 +49,8 @@ function MembersList(props) {
 		return <h2>Loading...</h2>;
 	} else {
 
+
+		console.log(props.teamMembers)
 		return (
 			<Content>
 				<p>Members({props.teamMembers.length})</p>
